🚚 chore(curation): `CurationGraph` 패키지 이동 및 키워드 기반 큐레이션 내비게이션 추가
- **패키지 구조 변경**
    - `CurationGraph.kt` 위치를 `navigation` 하위 패키지로 이동 (`com.linku.curation` → `com.linku.curation.navigation`)

- **내비게이션 로직 개선**
    - 반복되는 홈 이동 로직을 `goHome` 람다로 추출하여 코드 중복 제거 및 가독성 향상
    - `curation_card2` (키워드 상세) 및 `curation_keyword_links/{keyword}` (키워드별 링크 목록) 경로 추가
    - `curation_keyword_links` 이동 시 `Uri.encode/decode`를 사용하여 특수문자가 포함된 키워드 파라미터 전달 대응

- **UI 연결성 강화**
    - `CurationScreen`에 `onKeywordDetailClick` 콜백을 추가하여 키워드 상세 화면으로의 전환 기능 연결
    - `CurationKeywordDetailScreen` 및 `CurationKeywordLinksScreen` 스크린 컴포넌트 신규 연결 및 `nickname` 정보 주입 로직 반영
